% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLL.R
\name{computeLL}
\alias{computeLL}
\title{WP method background function computeLL}
\usage{
computeLL(p, NT, R0)
}
\arguments{
\item{p}{discretized version of the serial distribution}

\item{NT}{vector of case counts}

\item{R0}{basic reproductive ratio}
}
\value{
The function returns the variable \code{LL} which is the log-likelihood at the input variables and parameters.
}
\description{
This is a background/internal function called by \code{WP}.  It computes the log-likelihood.
}
