% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVIDCanadaPT}
\alias{COVIDCanadaPT}
\title{COVID-19 Canada Provincial and Territorial Case Counts, 2020-2023}
\format{
A data frame with 16799 observations on 4 variables:
\describe{
\item{region}{The two-letter code for the Canadian province or territory.}
\item{date}{The date of reporting in YYYY-MM-DD format.}
\item{cases}{The daily number of cases.}
\item{cumulative_cases}{The cumulative number of cases.}
}
}
\source{
\url{https://github.com/ccodwg/CovidTimelineCanada}
}
\usage{
COVIDCanadaPT
}
\description{
Daily COVID-19 case counts for each Canadian province and territory, from the
start of the pandemic until the end of 2023. Retrieved from the the COVID-19
Canada Open Data Working Group on 2024-05-11.
}
\keyword{datasets}
