% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID.R
\name{ID}
\alias{ID}
\title{ID method}
\usage{
ID(NT, mu)
}
\arguments{
\item{NT}{Vector of case counts.}

\item{mu}{Mean of the serial distribution. This needs to match case counts
in time units. For example, if case counts are weekly and the
serial distribution has a mean of seven days, then \code{mu} should
be set to one. If case counts are daily and the serial distribution
has a mean of seven days, then \code{mu} should be set to seven.}
}
\value{
\code{ID} returns a single value, the estimate of R0.
}
\description{
This function implements a least squares estimation method of R0 due to
Fisman et al. (PloS One, 2013). See details for implementation notes.
}
\details{
The method is based on a straightforward incidence decay model. The estimate
of R0 is the value which minimizes the sum of squares between observed case
counts and cases counts 'expected' under the model.

This method is based on an approximation of the SIR model, which is most
valid at the beginning of an epidemic. The method assumes that the mean of
the serial distribution (sometimes called the serial interval) is known. The
final estimate can be quite sensitive to this value, so sensitivity testing
is strongly recommended. Users should be careful about units of time (e.g.,
are counts observed daily or weekly?) when implementing.
}
\examples{
# Weekly data:
NT <- c(1, 4, 10, 5, 3, 4, 19, 3, 3, 14, 4)

# Obtain R0 when the serial distribution has a mean of five days.
ID(NT, mu = 5 / 7)

# Obtain R0 when the serial distribution has a mean of three days.
ID(NT, mu = 3 / 7)

}
