% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnaught-package.R
\docType{package}
\name{Rnaught-package}
\alias{Rnaught}
\alias{Rnaught-package}
\title{Rnaught: Estimate the Basic Reproduction Number (R0) of Infectious Diseases}
\description{
An R package and web application for estimating the basic reproduction number (R0) of infectious diseases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://MI2YorkU.github.io/Rnaught}
  \item \url{https://github.com/MI2YorkU/Rnaught}
  \item Report bugs at \url{https://github.com/MI2YorkU/Rnaught/issues}
}

}
\author{
\strong{Maintainer}: Naeem Model \email{me@nmode.ca}

Authors:
\itemize{
  \item Sawitree Boonpatcharanon
  \item Jane Heffernan
  \item Hanna Jankowski
}

Other contributors:
\itemize{
  \item Tatiana Krikella [contributor]
}

}
\keyword{internal}
