% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web}
\alias{web}
\title{Launch the Rnaught Web Application}
\usage{
web()
}
\value{
A Shiny app object for the Rnaught web application.
}
\description{
This is the entry point of the Rnaught web application, which creates and
returns a Shiny app object. When invoked directly, the web application is
launched.
}
\details{
The following dependencies are required to run the application:
\itemize{
\item \href{https://shiny.posit.co}{shiny}
\item \href{https://rstudio.github.io/bslib/}{bslib}
\item \href{https://rstudio.github.io/DT/}{DT}
\item \href{https://plotly-r.com}{plotly}
}

If any of the above packages are missing during launch, a prompt will appear
to install them.

To configure settings such as the port, host or default browser, set Shiny's
global options (see \code{\link[shiny:runApp]{shiny::runApp()}}).
}
