% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLL.R
\name{computeLL}
\alias{computeLL}
\title{WP method background function computeLL}
\usage{
computeLL(p, NT, R0)
}
\arguments{
\item{p}{Discretized version of the serial distribution.}

\item{NT}{Vector of case counts.}

\item{R0}{Basic reproductive ratio.}
}
\value{
This function returns the log-likelihood at the input variables and parameters.
}
\description{
This is a background/internal function called by \code{WP}.  It computes the log-likelihood.
}
\keyword{internal}
