% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVIDCanada}
\alias{COVIDCanada}
\title{COVID-19 Canada National Case Counts, 2020-2022}
\format{
A data frame with 1082 observations on 3 variables:
\describe{
  \item{date}{The date of reporting in YYYY-MM-DD format.}
  \item{cases}{The daily number of cases.}
  \item{cumulative_cases}{The cumulative number of cases.}
}
}
\source{
\url{https://opencovid.ca}
}
\usage{
COVIDCanada
}
\description{
Daily national COVID-19 case counts in Canada, from the start of the pandemic
until the end of 2022. Provided by the COVID-19 Canada Open Data Working
Group.
}
\keyword{datasets}
