% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WP_known.R
\name{WP_known}
\alias{WP_known}
\title{WP method background function WP_known}
\usage{
WP_known(NT, p)
}
\arguments{
\item{NT}{Vector of case counts.}

\item{p}{Discretized version of the serial distribution.}
}
\value{
The function returns the maximum likelihood estimator of R0.
}
\description{
This is a background/internal function called by \code{WP}. It computes the maximum
likelihood estimator of R0 assuming that the serial distribution is known and finite discrete.
}
